+++ CHANGES.md
@@ -6,7 +6,9 @@
 
 -   Fix writing non-string object columns with arrow (#630).
 -   Fix writing empty string category columns with arrow fails (#621).
+-   Fix Time type columns being skipped with `use_arrow=False` (#617).
 -   Fix overwriting a corrupt fileGDB directory (#600).
+-   Fix attribute data being incorrectly written with KML driver (#650).
 
 ## 0.12.1 (2025-11-28)